Write functions that process JSON data from API
I wrote three functions that get the info I need: a function for
today's weather, forecasted weather for the next 22 hours, and
forecasted weather for the next 7 days. These functions make use of
async functions to handle the asynchronous code, however in a future
commit, I will also convert these functions to promise/then functions
for practice. These functions pull the weather data from the weatherapi,
which is free to make an account online and get access to an API key.