Set up input form and event handler to return data
In the previous commit, Each function that processes the weather api
data did not return anything, they only logged to the console. I changed
this so that the functions return a promises with objects and arrays of
objects as the resolved data. These objects contain all of the required,
processed data. I also created an event handler function that invokes
all of the data processing functions, and logs to the console an array
of this data once all promises are resolved. Finally, I added a simple
input form in the html, so that the user can search for a city and its
weather.