Add functionality for changing units
Now when the promises are resolved in the flow control function, the
results array gets copied to a global array called locationInfo. Now
thanks to an event listener on the today's temperature and a new
function for changing the units, the info with different units get
pulled from the locationInfo array. The units that are pulled from the
array depend on a new boolean variable named isCelsius.