Fix position issue with block-4 of today's weather
Now the fourth block in the today's weather section is always positioned
in line with the content at the most right of the content-wrapper, and
is always visible with a fully widened window. This section, like
before, will become hidden as the window slowly shrinks horizontally.
Finally, there were a couple of weather conditions (blowing snow and
blizzard) that were not originally included in the .js file which have
now been included.