Make changing of units into a separate function
Whatever text content being appended to the dom that has units that may
be changed at the users request is now appended to the dom by a separate
js function called appendTemperatureAndWindSpeed. The original js
function that used to append everything to the dom has now been renamed
to appendMainContentAndStyle, because it no longer performs the job that
the new appendTemperatureAndWindSpeed function does. Now, when the user
clicks on the todays temperature to change the units used in the app,
only the necessary text content is replaces, rather than the entire
dom tree, which makes this action look much smoother.

Finally, I cleaned up the html file to not include the original text
content from the conception of this project, and I removed the
unnecessary comments from the .css file.