Add footer with credits
The only issue with this is that the link to the odin project is
un-clickable. I believe this is because the credits are set at a z-index
of -1. If I don't do this however, the loading screen does not darken
the credits for some reason. This will have to be worked on as a future
improvement.