Fix day issue on Safari browser
On the Safari browser, the current day was showing as invalid. This is
because the getWeatherToday function was creating a new date from the
API pulled date, which was not in the standard form of 'YYYY-MM-DDTHH:mm
:ss.sssZ'. The google chrome and firefox browsers were able to work with
the non standard date, but Safari was not. I updated the function to
convert the non-standard date to standard, and now it works on Safari.